refactor: Editor to functional component & has been moved to a separate folder
